+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # OUDS iOS library changelog
 
 All notable changes to this project will be documented in this file.
@@ -25,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GitHub Actions `checkout` from v5 to v6 for all workflows
 - Gem `fastlane` gem from v2.228.0 to v2.229.1
 - `Wireframe` color raw tokens names
+- Update `switch` components to v1.5.0 (Orange-OpenSource/ouds-ios#1138)
+- Update `radio` components to v1.4.0 (Orange-OpenSource/ouds-ios#1139)
 - Update `checkbox` components to v2.4.0 (Orange-OpenSource/ouds-ios#1137)
 - `Wireframe` theme `border` semantic tokens (tokens libraries System v2.3) (Orange-OpenSource/ouds-ios#1158)
 - `size` semantic tokens (tokens libraries System v2.3) (Orange-OpenSource/ouds-ios#1158)